Checkpoint commit for RTR gleaning.
Add some new functionality to an RTR. When configured, an RTR can glean
mappings for packets it decapsulates. This allows for an xTR to not run
a LISP control-plane. The xTR simply encapsulates packets to configured RTRs.
Return packets will arrive to an xTR since the RTR will glean the source EID
and source RLOC from decapsulated packets. A new command "lisp glean-
mapping" is required on the RTR. When configured, the RTR will add the
source-EID and source-RLOC to the map-cache. See lisp.config.example
for a description of the command and some examples. EID-prefixes are used
to permit ranges of EIDs allowed to be gleaned as well are RLOC-prefixes
so RLOCs can be controlled. A single or range of instance-IDs can also be
configured.