(1) Move lisp_igmp_process_packets() to lisp.py to be used by both the lisp-rtr and lisp-etr     processes. (2) Have the caller of lisp_igmp_process_packets() send Map-Register mesasges if register     entries are returned by processing all (S,G) and (*,G) entriess in an IGMP report. (3) Glean source RLOC information for IGMP sender in the lisp-rtr process. (4) Ignore Map-Notify messages for gleaned multicast map-cache entries.
Left to do:

(1) Time out RLE entries when no IGMP reports received.
(2) We need to IGMP query each RLE entry so it can respond with a report.
(3) A gleaned entry should override a mapping system entry. There can be a race condition
    where sources are sending before gleaned members are learned. The map-cache would
    have a set of RLEs of all RTRs but what it needs is a list of LISP-MNs.
(4) When the source RLOC information changes for an EID, see if the EID is joined to any
    multicast entry and make the change sooner rather than later (up to 1 minute) for the
    next IGMP report to be sent.